nba_lineup_rag/requirements.txt
# Vector Database & Embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0
torch>=2.1.0

# Data Processing
numpy>=1.26.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
readability-lxml>=0.8.1

# RSS Parsing
feedparser>=6.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
rapidfuzz>=3.9.0

# LLM & LangChain
openai>=1.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langchain-community>=0.3.0
ipywidgets>=8.0.0

